requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.32.4
langchain==0.3.26
openai==1.86.0
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langchain-community==0.3.26
faiss-cpu==1.8.0  # optional, will add pgvector later
psycopg2-binary==2.9.9
pgvector==0.2.4
prometheus_fastapi_instrumentator==5.9.1
pypdf==4.2.0
reportlab==4.2.0
matplotlib==3.9.0